refactor(zoom): to use button-group and switch to features dir
