+++ src/features/index.ts
@@ -1,1 +1,2 @@
 export * from "./toolbar";
+export * from "./zoom";
